Fixed a warning in dx11.pipeline.cpp. Split dx11.pipeline.cpp into two. Removed pragma comment lib entries specifying .debug.lib files. DX11.Renderer now included in sharpmake builds.

+++ dx11.renderer/dx11.cube.texture.cpp
@@ -175,7 +175,7 @@ namespace Rococo::DX11
 				D3D11_TEXTURE2D_DESC desc;
 				cubeTexture->GetDesc(&desc);
 
-				if (desc.Width != materialArray.Width())
+				if (desc.Width != (uint32) materialArray.Width())
 				{
 					cubeTexture = nullptr;
 					cubeTextureView = nullptr;
